Fix review issues, unify naming
